# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AugmentedQuill",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_name}/checkpoints": {
      "get": {
        "tags": ["Checkpoints"],
        "summary": "Api Get Checkpoints",
        "description": "Handle the API request to get checkpoints.",
        "operationId": "api_get_checkpoints_api_v1_projects__project_name__checkpoints_get",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_name}/checkpoints/create": {
      "post": {
        "tags": ["Checkpoints"],
        "summary": "Api Create Checkpoint",
        "description": "Handle the API request to create checkpoint.",
        "operationId": "api_create_checkpoint_api_v1_projects__project_name__checkpoints_create_post",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_name}/checkpoints/load": {
      "post": {
        "tags": ["Checkpoints"],
        "summary": "Api Load Checkpoint",
        "description": "Handle the API request to load checkpoint.",
        "operationId": "api_load_checkpoint_api_v1_projects__project_name__checkpoints_load_post",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointLoadDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_name}/checkpoints/delete": {
      "post": {
        "tags": ["Checkpoints"],
        "summary": "Api Delete Checkpoint",
        "description": "Handle the API request to delete checkpoint.",
        "operationId": "api_delete_checkpoint_api_v1_projects__project_name__checkpoints_delete_post",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointLoadDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckpointInfo": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": ["timestamp"],
        "title": "CheckpointInfo",
        "description": "Represents the CheckpointInfo type."
      },
      "CheckpointListResponse": {
        "properties": {
          "checkpoints": {
            "items": {
              "$ref": "#/components/schemas/CheckpointInfo"
            },
            "type": "array",
            "title": "Checkpoints"
          }
        },
        "type": "object",
        "required": ["checkpoints"],
        "title": "CheckpointListResponse",
        "description": "Represents the CheckpointListResponse type."
      },
      "CheckpointLoadDeleteRequest": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": ["timestamp"],
        "title": "CheckpointLoadDeleteRequest",
        "description": "Represents the CheckpointLoadDeleteRequest type."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}